regractor: suggested review changes
